Create TravelPayStatusContent and reorganize TravelPayStatusApp
